refactor notes

+++ Sagas/ListSagas.js
@@ -1,4 +1,4 @@
-import DataActions from "../Redux/DataRedux";
+import ListActions from "../Redux/ListRedux";
 import { call, put } from "redux-saga/effects";
 import Utils from "../Utils";
 
@@ -6,9 +6,9 @@ export function* request(api, { name }) {
   var json = yield api.get(`${Utils.config.url}/api/db/${name}`);
 
   if (json.error) {
-    yield put(DataActions.dataFailure(json.error));
+    yield put(ListActions.listFailure(name, json.error));
   } else {
-    yield put(DataActions.dataSuccess(name, json));
+    yield put(ListActions.listSuccess(name, json));
   }
 }
 
@@ -19,9 +19,9 @@ export function* add(api, { name, item, duplicate }) {
   });
 
   if (json.error) {
-    yield put(DataActions.dataFailure(json.error));
+    yield put(ListActions.listFailure(json.error));
   } else {
-    yield put(DataActions.dataAddSuccess(name, json));
+    yield put(ListActions.listAddSuccess(name, json));
   }
 }
 
@@ -29,21 +29,25 @@ export function* addDetail(api, { name, id, item }) {
   var json = yield api.post(`${Utils.config.url}/api/db/${name}/${id}`, item);
 
   if (json.error) {
-    yield put(DataActions.dataFailure(json.error));
+    yield put(ListActions.listFailure(json.error));
   } else {
-    yield put(DataActions.dataAddSuccess(name, json));
+    yield put(ListActions.listAddSuccess(name, json));
   }
 }
 
 export function* remove(api, { name, id }) {
-  var json = yield api.delete(
-    `${Utils.config.url}/api/db/${name}/apiId/${encodeURI(id)}`
-  );
+  if (id) {
+    var json = yield api.delete(
+      `${Utils.config.url}/api/db/${name}/apiId/${encodeURI(id)}`
+    );
 
-  if (json.error) {
-    yield put(DataActions.dataFailure(json.error));
-  } else {
-    yield put(DataActions.dataSuccess(name, json));
+    if (json.error) {
+      yield put(ListActions.listFailure(json.error));
+    } else {
+      yield put(ListActions.listSuccess(name, []));
+    }
+  }else {
+    yield put(ListActions.listFailure("no apiId"));
   }
 }
 
@@ -51,8 +55,8 @@ export function* removeAll(api, { name }) {
   var json = yield api.delete(`${Utils.config.url}/api/db/${name}`);
 
   if (json.error) {
-    yield put(DataActions.dataFailure(json.error));
+    yield put(ListActions.listFailure(json.error));
   } else {
-    yield put(DataActions.dataSuccess(name, []));
+    yield put(ListActions.listSuccess(name, []));
   }
 }
